Crate jail

A rust library for FreeBSD jails.

it aims to provide the features exposed by the FreeBSD Jail Library jail(3)

Modules

param
Module for inspection and manipulation of jail parameters

process
Jail-Specific extensions to the std::process module

Structs

RunningJail
Represents a running jail.

RunningJailIter
An Iterator over running Jails

StoppedJail
Represent a stopped jail including all information required to start it

Enums

Jail
Represents a running or stopped jail.

JailError
An enum for error types of the Jail.