Trait jail::process::Jailed

pub trait Jailed {
    fn jail(&mut self, jail: &RunningJail) -> &mut Command;
}

Extension to the std::process::Command builder to run the command in a jail.

Adds a before_exec hook to the std::process::Command builder that calls the jail_attach(2) syscall.

Examples

use jail::process::Jailed;

let output = Command::new("/hostname")
             .jail(&jail)
             .output()
             .expect("Failed to execute command");

println!("output: {:?}", output.stdout);

Required methods

fn jail(&mut self, jail: &RunningJail) -> &mut Command

Sets the child process to be executed within a jail. This translates to calling jail_attach in the child process. Failure in the jail_attach call will cause the spawn to fail.

Loading content...