Struct jail::RunningJail

pub struct RunningJail {
    pub jid: i32,
}

Represents a running jail.

Fields
jid: i32

The jid of the jail

Methods

impl RunningJail

Represent a running jail.

pub fn from_jid(jid: i32) -> RunningJail

Create a RunningJail instance given a jid.

No checks will be performed.

Examples

use jail::RunningJail;

let running = RunningJail::from_jid(jid);

pub fn from_name(name: &str) -> Result<RunningJail, JailError>

Create a RunningJail given the jail name.

The jid will be internally resolved using jail_getid.

Examples

use jail::RunningJail;

let running = RunningJail::from_name("testjail_from_name")
    .expect("Could not get testjail");

pub fn name(self: &RunningJail) -> Result<String, JailError>

Return the jail's name.

Examples

assert_eq!(running.name().unwrap(), "testjail_name");

pub fn path(self: &RunningJail) -> Result<PathBuf, JailError>

Return the jail's path.

Examples

let path = running.path()
    .expect("Could not get path");

pub fn hostname(self: &RunningJail) -> Result<String, JailError>

Return the jail's name.

The name will be internall resolved using jail_getname.

Examples

assert_eq!(running.hostname().unwrap(), "testjail.example.com");

pub fn ips(self: &RunningJail) -> Result<Vec<IpAddr>, JailError>

Get the IP addresses

Examples

let ips = running.ips()
    .expect("could not get ip addresses");
assert_eq!(ips[0], "127.0.1.2".parse::<IpAddr>().unwrap());
assert_eq!(ips[1], "fe80::2".parse::<IpAddr>().unwrap());

pub fn param(&self, name: &str) -> Result<Value, JailError>

Return a jail parameter.

Examples

let hostuuid = running.param("host.hostuuid")
    .expect("could not get jail hostuuid");

pub fn params(&self) -> Result<HashMap<String, Value>, JailError>

Return a HashMap of all jail parameters.

Examples

use jail::param;

let params = running.params()
    .expect("could not get all parameters");

assert_eq!(
    params.get("allow.raw_sockets"),
    Some(&param::Value::Int(1))
);

pub fn param_set(&self, name: &str, value: Value) -> Result<(), JailError>

Set a jail parameter.

Examples

use jail::param;
running.param_set("allow.raw_sockets", param::Value::Int(1))
    .expect("could not set parameter");

pub fn kill(self: RunningJail) -> Result<(), JailError>

Kill a running jail, consuming it.

This will kill all processes belonging to the jail, and remove any children of that jail.

Examples

running.kill();

pub fn save(self: &RunningJail) -> Result<StoppedJail, JailError>

Create a StoppedJail from a RunningJail, while not consuming the RunningJail.

This can be used to clone the config from a RunningJail.

Examples

let stopped = running
    .save()
    .expect("could not save jail configuration");

assert_eq!(stopped.name, Some("testjail_save".into()));
assert_eq!(stopped.hostname, Some("testjail_save.example.com".into()));

pub fn stop(self: RunningJail) -> Result<StoppedJail, JailError>

Stop a jail, keeping its configuration in a StoppedJail.

This is a wrapper around save and kill.

Examples

let stopped = running
    .stop()
    .expect("failed to stop jail");

//assert_eq!(stopped.name, Some("testjail_save".into()));
//assert_eq!(stopped.hostname, Some("testjail_save.example.com".into()));

pub fn restart(self: RunningJail) -> Result<RunningJail, JailError>

Restart a jail by stopping it and starting it again

This is a wrapper around RunningJail::stop and StoppedJail::start

Examples

let old_jid = running.jid;
let running = running.restart()
    .expect("failed to restart jail");
assert!(running.jid != old_jid);

pub fn all() -> RunningJails

Returns an Iterator over all running jails on this host.

Examples

use jail::RunningJail;

for running in RunningJail::all() {
    println!("jail: {}", running.name().unwrap());
}