Struct jail::StoppedJail

pub struct StoppedJail {
    pub path: Option<PathBuf>,
    pub name: Option<String>,
    pub hostname: Option<String>,
    pub params: HashMap<String, Value>,
    pub ips: Vec<IpAddr>,
}

Represent a stopped jail including all information required to start it

Fields
path: Option<PathBuf>

The path of root file system of the jail

name: Option<String>

The jail name

hostname: Option<String>

The jail hostname

params: HashMap<String, Value>

A hashmap of jail parameters and their values

ips: Vec<IpAddr>

A list of IP (v4 and v6) addresses to be assigned to this jail

Methods

impl StoppedJail

pub fn new<P: Into<PathBuf>>(path: P) -> StoppedJail

Create a new Jail instance given a path.

Examples

use jail::StoppedJail;

let j = StoppedJail::new("/rescue");

pub fn start(self: StoppedJail) -> Result<RunningJail, JailError>

Start the jail

This will call jail_create internally. This will consume the StoppedJail and return a Result<RunningJail,Error>.

Examples

use jail::StoppedJail;

let stopped = StoppedJail::new("/rescue");
let running = stopped.start().unwrap();

pub fn name<S: Into<String>>(self, name: S) -> Self

Set the jail name

Examples

let mut stopped = StoppedJail::new("/rescue")
    .name("test_stopped_name");

assert_eq!(stopped.name, Some("test_stopped_name".to_string()));

pub fn hostname<S: Into<String>>(self, hostname: S) -> Self

Set the jail name

Examples

let mut stopped = StoppedJail::new("/rescue")
    .hostname("example.com");

assert_eq!(stopped.hostname, Some("example.com".to_string()));

pub fn param<S: Into<String>>(self, param: S, value: Value) -> Self

Set a jail parameter

Examples

use jail::param;

let mut stopped = StoppedJail::new("/rescue")
    .param("allow.raw_sockets", param::Value::Int(1));

pub fn ip(self, ip: IpAddr) -> Self

Add an IP Address

Examples

let mut stopped = StoppedJail::new("rescue")
    .ip("127.0.1.1".parse().expect("could not parse 127.0.1.1"))
    .ip("fe80::2".parse().expect("could not parse ::1"));